moved general instructions to shared file
