+++ shared.h
@@ -1,0 +1,3 @@
+
+char getOpCode(char const *instruction);
+char identifyOpCode(char opCode);
